Some restructuring of gnimag target
