--- gnimag/Sources/Products/gnimag/Sources/IO/Android/README.md
@@ -1,7 +1,0 @@
-## Android
-
-Scrcpy and Vysor are applications that can stream your Android device onto your Mac desktop. They can also send input events to the Android device. This makes them perfect for our use-case.
-
-Downloads:
-- [scrcpy](https://github.com/Genymobile/scrcpy)
-- [vysor](https://vysor.io)
