+++ gnimag/Sources/Products/gnimag/Sources/Screen-IO/Components/MouseTapper.swift
