+++ gnimag/Sources/Products/gnimag/Sources/Screen-IO/NoopTapper.swift
