+++ gnimag/Sources/Products/gnimag/Sources/Screen-IO/WindowInteractor.swift
